birras/requirements.txt
Django==3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
requests
drf-yasg
gunicorn
whitenoise
celery[redis]
flower
django-grappelli
django-prometheus
psycopg2-binary==2.8.5
jinja2